pyproject.toml
[tool.poetry]
name = "eazy-quizzy"
version = "0.1.0"
description = ""
authors = ["Ruslan Atarov <ruslanatarov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.7"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
pre-commit = "^3.5.0"
gunicorn = "^21.2.0"
django-nested-admin = "^4.0.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
black = "^23.11.0"
isort = "^5.12.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py311']
exclude = 'migrations'
line_length = 79

[tool.isort]
line_length = 79
skip = ['migrations', 'settings']
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true

[tool.ruff]
target-version = "py311"
exclude = [
    "*migrations/",
    "settings.py",
]
line-length = 79